Split transactiion planner
