+++ app/frontend/types.ts
@@ -1,7 +1,7 @@
 import {TxSigned, TxAux, CborizedCliWitness} from './wallet/shelley/types'
 import {CaTxEntry, NextRewardDetail, RewardType, TokenObject} from './wallet/backend-types'
 import {Network, WalletName} from './wallet/types'
-import {TxPlan} from './wallet/shelley/shelley-transaction-planner'
+import {TxPlan} from './wallet/shelley/transaction'
 import {_UnsignedTxParsed} from './helpers/cliParser/types'
 
 export type BIP32Path = number[]
