+++ app/frontend/wallet/shelley/transaction/index.ts
@@ -1,0 +1,3 @@
+export {selectMinimalTxPlan} from './shelley-transaction-planner'
+export * from './types'
+export {computeMinUTxOLovelaceAmount, computeRequiredTxFee} from './utils'
